+++ nes-runtime/src/Execution/Operators/Scan.cpp
@@ -35,7 +35,7 @@ void Scan::open(ExecutionContext& ctx, RecordBuffer& recordBuffer) const {
     // iterate over records in buffer
     auto numberOfRecords = recordBuffer.getNumRecords();
     auto bufferAddress = recordBuffer.getBuffer();
-    for (Value<UInt64> i = (uint64_t) 0; i < numberOfRecords; i = i + (uint64_t) 1) {
+    for (Value<UInt64> i = 0_u64; i < numberOfRecords; i = i + 1_u64) {
         auto record = memoryProvider->read(projections, bufferAddress, i);
         child->execute(ctx, record);
     }
