+++ nes-runtime/src/Experimental/Interpreter/Operators/Aggregation/AvgFunction.cpp
@@ -32,7 +32,7 @@ std::unique_ptr<AggregationState> AvgFunction::createState() {
                     return std::make_unique<AvgState>(Value<>((uint32_t) 0), Value<>((int64_t) 0));
                 };
                 case Nautilus::IR::Types::IntegerStamp::I64: {
-                    return std::make_unique<AvgState>(Value<>((uint64_t) 0), Value<>((int64_t) 0));
+                    return std::make_unique<AvgState>(Value<>(0_u64), Value<>((int64_t) 0));
                 };
             }
         } else {
@@ -87,7 +87,7 @@ Value<Any> AvgFunction::lower(std::unique_ptr<AggregationState>& state) {
 }
 std::unique_ptr<AggregationState> AvgFunction::loadState(Value<MemRef>& ref) {
     auto counter = ref.load<Int64>();
-    auto valueRef = (ref + (uint64_t) 8).as<MemRef>();
+    auto valueRef = (ref + 8_u64).as<MemRef>();
     if (auto integerStamp = cast_if<IR::Types::IntegerStamp>(stamp.get())) {
         if (integerStamp->isUnsigned()) {
             switch (integerStamp->getBitWidth()) {
@@ -135,7 +135,7 @@ std::unique_ptr<AggregationState> AvgFunction::loadState(Value<MemRef>& ref) {
 void AvgFunction::storeState(Value<MemRef>& ref, std::unique_ptr<AggregationState>& state) {
     auto sumState = (AvgState*) state.get();
     ref.store(sumState->count);
-    auto valueRef = (ref + (uint64_t) 8).as<MemRef>();
+    auto valueRef = (ref + 8_u64).as<MemRef>();
     valueRef.store(sumState->sum);
 }
 uint64_t AvgFunction::getStateSize() const { return 16; }
