+++ nes-runtime/tests/UnitTests/Execution/Aggregation/AggregationFunctionDataTypeTest.cpp
@@ -77,7 +77,7 @@ class AggregationFunctionDataTypeTest : public Testing::NESBaseTest, public ::te
         } else if (std::equal(dataTypeString.begin(), dataTypeString.end(), "ui32")) {
             return Nautilus::Value<Nautilus::UInt32>((uint32_t) 1);
         } else if (std::equal(dataTypeString.begin(), dataTypeString.end(), "ui64")) {
-            return Nautilus::Value<Nautilus::UInt64>((uint64_t) 1);
+            return Nautilus::Value<Nautilus::UInt64>(1_u64);
         } else if (std::equal(dataTypeString.begin(), dataTypeString.end(), "f32")) {
             return Nautilus::Value<Nautilus::Float>((float) 1.0);
         } else if (std::equal(dataTypeString.begin(), dataTypeString.end(), "f64")) {
